# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SHOP Quote API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalQuotes: number;
  draftQuotes: number;
  sentQuotes: number;
  wonQuotes: number;
  lostQuotes: number;
  expiredQuotes: number;
  totalQuotedValue: number;
  wonValue: number;
  conversionRate: number;
  followUpNeeded: number;
}

export interface Customer {
  id: number;
  companyName: string;
  contactName: string;
  email: string;
  phone: string;
  address: string;
  notes: string;
  createdAt: string;
}

export interface CustomerInput {
  companyName: string;
  contactName: string;
  email: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface CustomerUpdate {
  companyName?: string;
  contactName?: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface Machine {
  id: number;
  name: string;
  machineType: string;
  axisCount: number;
  hourlyRate: number;
  setupRate: number;
  notes: string;
  active: boolean;
  createdAt: string;
}

export interface MachineInput {
  name: string;
  machineType: string;
  axisCount?: number;
  hourlyRate: number;
  setupRate: number;
  notes?: string;
  active?: boolean;
}

export interface MachineUpdate {
  name?: string;
  machineType?: string;
  axisCount?: number;
  hourlyRate?: number;
  setupRate?: number;
  notes?: string;
  active?: boolean;
}

export interface QuoteLineItem {
  id: number;
  quoteId: number;
  partName: string;
  drawingNumber: string;
  revision: string;
  quantity: number;
  material: string;
  processType: string;
  /** @nullable */
  machineId: number | null;
  toleranceClass: string;
  surfaceFinish: string;
  lineItemType: string;
  hiddenFromPdf: boolean;
  complexity: string;
  setupHours: number;
  programmingHours: number;
  machiningMinutesPerPart: number;
  inspectionHours: number;
  deburringMinutesPerPart: number;
  materialCostPerUnit: number;
  materialWastagePercentage: number;
  toolingAllowance: number;
  outsideProcessing: number;
  packaging: number;
  delivery: number;
  riskPercentage: number;
  profitMarginPercentage: number;
  discountPercentage: number;
  vatEnabled: boolean;
  vatRate: number;
  machineHourlyRate: number;
  machineSetupRate: number;
  setupCost: number;
  programmingCost: number;
  machiningCost: number;
  inspectionCost: number;
  deburringCost: number;
  materialCostTotal: number;
  directCost: number;
  riskValue: number;
  costBeforeMargin: number;
  sellPrice: number;
  pricePerPart: number;
  vatAmount: number;
  totalIncVat: number;
  toolingRecommendation: string;
  materialRecommendation: string;
  coolantRecommendation: string;
}

export interface QuoteLineItemInput {
  partName: string;
  drawingNumber?: string;
  revision?: string;
  quantity: number;
  material: string;
  processType: string;
  /** @nullable */
  machineId?: number | null;
  toleranceClass?: string;
  surfaceFinish?: string;
  complexity?: string;
  lineItemType?: string;
  hiddenFromPdf?: boolean;
  setupHours: number;
  programmingHours: number;
  machiningMinutesPerPart: number;
  inspectionHours: number;
  deburringMinutesPerPart: number;
  materialCostPerUnit: number;
  materialWastagePercentage: number;
  toolingAllowance: number;
  outsideProcessing: number;
  packaging: number;
  delivery: number;
  riskPercentage: number;
  profitMarginPercentage: number;
  discountPercentage?: number;
  vatEnabled?: boolean;
  vatRate?: number;
  toolingRecommendation?: string;
  materialRecommendation?: string;
  coolantRecommendation?: string;
}

export interface QuoteSummary {
  id: number;
  quoteNumber: string;
  customerId: number;
  customerName: string;
  status: string;
  quoteDate: string;
  validUntil: string;
  totalValue: number;
  createdAt: string;
}

export interface Quote {
  id: number;
  quoteNumber: string;
  customerId: number;
  status: string;
  lostReason: string;
  quoteDate: string;
  validUntil: string;
  quoteRevision: string;
  revisionNotes: string;
  leadTime: string;
  deliveryTerms: string;
  notes: string;
  internalNotes: string;
  paymentTerms: string;
  termsAndConditions: string;
  materialCertIncluded: boolean;
  inspectionReportIncluded: boolean;
  fairIncluded: boolean;
  cmmReportIncluded: boolean;
  priceBreakQtys: string;
  lineItems: QuoteLineItem[];
  createdAt: string;
  updatedAt: string;
}

export interface QuoteInput {
  customerId: number;
  status?: string;
  lostReason?: string;
  quoteDate?: string;
  validUntil?: string;
  quoteRevision?: string;
  revisionNotes?: string;
  leadTime?: string;
  deliveryTerms?: string;
  notes?: string;
  internalNotes?: string;
  paymentTerms?: string;
  termsAndConditions?: string;
  materialCertIncluded?: boolean;
  inspectionReportIncluded?: boolean;
  fairIncluded?: boolean;
  cmmReportIncluded?: boolean;
  priceBreakQtys?: string;
  lineItems: QuoteLineItemInput[];
}

export interface QuoteUpdate {
  customerId?: number;
  status?: string;
  lostReason?: string;
  quoteDate?: string;
  validUntil?: string;
  quoteRevision?: string;
  revisionNotes?: string;
  leadTime?: string;
  deliveryTerms?: string;
  notes?: string;
  internalNotes?: string;
  paymentTerms?: string;
  termsAndConditions?: string;
  materialCertIncluded?: boolean;
  inspectionReportIncluded?: boolean;
  fairIncluded?: boolean;
  cmmReportIncluded?: boolean;
  priceBreakQtys?: string;
  lineItems?: QuoteLineItemInput[];
}

export interface Settings {
  id: number;
  companyName: string;
  address: string;
  email: string;
  phone: string;
  website: string;
  vatNumber: string;
  currency: string;
  logoUrl: string;
  bankName: string;
  accountName: string;
  accountNumber: string;
  sortCode: string;
  iban: string;
  swiftBic: string;
  showBankDetails: boolean;
  defaultHourlyRate: number;
  defaultSetupRate: number;
  defaultMarginPercentage: number;
  vatEnabled: boolean;
  vatRate: number;
  quoteValidityDays: number;
  defaultLeadTime: string;
  defaultDeliveryTerms: string;
  paymentTerms: string;
  termsAndConditions: string;
}

export interface SettingsUpdate {
  companyName?: string;
  address?: string;
  email?: string;
  phone?: string;
  website?: string;
  vatNumber?: string;
  currency?: string;
  logoUrl?: string;
  bankName?: string;
  accountName?: string;
  accountNumber?: string;
  sortCode?: string;
  iban?: string;
  swiftBic?: string;
  showBankDetails?: boolean;
  defaultHourlyRate?: number;
  defaultSetupRate?: number;
  defaultMarginPercentage?: number;
  vatEnabled?: boolean;
  vatRate?: number;
  quoteValidityDays?: number;
  defaultLeadTime?: string;
  defaultDeliveryTerms?: string;
  paymentTerms?: string;
  termsAndConditions?: string;
}

export interface QuoteDrawing {
  id: number;
  quoteId: number;
  filename: string;
  contentType: string;
  objectPath: string;
  drawingNumber: string;
  revision: string;
  uploadedAt: string;
}

export interface CreateQuoteDrawingBody {
  objectPath: string;
  filename: string;
  contentType: string;
  drawingNumber?: string;
  revision?: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResponseMetadata = {
  name?: string;
  size?: number;
  contentType?: string;
};

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: RequestUploadUrlResponseMetadata;
}