# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SHOP Quote API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /machines:
    get:
      operationId: listMachines
      tags: [machines]
      summary: List all machines
      responses:
        "200":
          description: List of machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Machine"
    post:
      operationId: createMachine
      tags: [machines]
      summary: Create a machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
  /machines/{id}:
    get:
      operationId: getMachine
      tags: [machines]
      summary: Get a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Machine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "404":
          description: Not found
    patch:
      operationId: updateMachine
      tags: [machines]
      summary: Update a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
    delete:
      operationId: deleteMachine
      tags: [machines]
      summary: Delete a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Customer:
      type: object
      required:
        - id
        - companyName
        - contactName
        - email
        - phone
        - address
        - notes
        - createdAt
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
        createdAt:
          type: string
    CustomerInput:
      type: object
      required:
        - companyName
        - contactName
        - email
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    Machine:
      type: object
      required:
        - id
        - name
        - machineType
        - axisCount
        - hourlyRate
        - setupRate
        - notes
        - active
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        machineType:
          type: string
        axisCount:
          type: integer
        hourlyRate:
          type: number
        setupRate:
          type: number
        notes:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
    MachineInput:
      type: object
      required:
        - name
        - machineType
        - hourlyRate
        - setupRate
      properties:
        name:
          type: string
        machineType:
          type: string
        axisCount:
          type: integer
        hourlyRate:
          type: number
        setupRate:
          type: number
        notes:
          type: string
        active:
          type: boolean
    MachineUpdate:
      type: object
      properties:
        name:
          type: string
        machineType:
          type: string
        axisCount:
          type: integer
        hourlyRate:
          type: number
        setupRate:
          type: number
        notes:
          type: string
        active:
          type: boolean
...